% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.reml_sketch}
\alias{plot.reml_sketch}
\title{Plot the path of a REML fit}
\usage{
\method{plot}{reml_sketch}(x, ...)
}
\arguments{
\item{x}{An object from \code{\link[=reml_sketch]{reml_sketch()}} or \code{\link[=reml_exact]{reml_exact()}}.}

\item{...}{Passed to \code{\link[graphics:matplot]{graphics::matplot()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Draws the genetic and residual variance estimates at each iteration, which
shows whether the fit settled or was still moving when it stopped.
}
\examples{
set.seed(1)
dat <- sim_genomic(n = 200, p = 400, h2 = 0.5)
plot(reml_exact(dat$y, dat$G))
}
